backend/data_ingestion_pipeline/requirements.txt
google-generativeai
python-dotenv
pymupdf
pymupdf4llm 